+++ src/ECS/mod.rs
@@ -19,6 +19,7 @@ pub mod misc;
 pub mod events;
 pub mod commands;
 pub mod prefab;
+pub mod fetch;
 
 pub mod prelude;
