+++ src/ECS/prelude.rs
@@ -11,4 +11,5 @@ pub use super::{comp::*,
     dispatcher::*,
     misc::*,
     events::*,
+    fetch::*
 };
